forward_gps_tracker/requirements.txt
gps-tracker
paho-mqtt